refactor / change css class sbs-newq to sbs-isolated
Instead of using the "sbs-newq" class we use the "sbs-isolated" class.
The only difference is the name. This will be used for more pages than
just the "New Question" page.

The "tag-editor-p" class is no longer used because it was redundant,
already covered by the "form-item" class.

The branching around the code which toggles the "sbs-isolated" class is
now more adaptable to other isolated editing pages.