refactor / feature / separate code for different isolated editing pages, add sbs for answering own question
The code for the various non-/answers/#/ pages is now split into
different branches of a conditional to make it more organized and easier
to add support for any pages I've missed in the future.

This update also adds sbs mode for the editor that appears when you
click the "Answer your own question" checkbox on the /ask/ page.